+++ lenstras_algo.py
